fixed wrong config used for acc experiments

+++ scripts/common.py
@@ -50,7 +50,7 @@ def parse_args():
     )
     parser.add_argument(
         "--max_tokens",
-        type=int,
+        type=float,
         default=32,
         help="Maximum number of thousand (in units of 1024) tokens (used in sampling parameter).",
     )
